DAY 9: Adds working combat system. Adds working stage system. Adds enemy strength progression by stage. Adds various needed scriptable objects. Adds panel for realtime enemy stats.

+++ games/no-name-yet [2 WEEKS]/Assets/ScriptableObjets/IntVariable.cs
@@ -7,10 +7,10 @@ using UnityEngine;
 [CreateAssetMenu]
 public class IntVariable : ScriptableObject, ISerializationCallbackReceiver
 {
-    public float InitialValue;
+    public int InitialValue;
 
     [NonSerialized]
-    public float RuntimeValue;
+    public int RuntimeValue;
 
     public void OnAfterDeserialize()
     {
